DTO 수정, 공지게시판 글쓰기 기능 구현

+++ finalproject6/src/main/java/com/kh/final6/entity/NoticeAttachDto.java
@@ -8,6 +8,6 @@ import lombok.NoArgsConstructor;
 
 @Data @NoArgsConstructor @AllArgsConstructor @Builder
 public class NoticeAttachDto {
-	private long noticeNo;
+	private int noticeNo;
 	private long attachmentNo;
 }
